fix: split proxy.go into platform-specific files for Windows build
ProxyCommand support uses Unix-only syscalls (Kill with negative PID,
SysProcAttr.Setpgid) that don't exist on Windows. Split into proxy.go
(shared), proxy_unix.go (syscall-based cleanup), and proxy_windows.go
(Process.Kill fallback) with build tags.
